Add sysusers file
This allows the polkitd user to be created by systemd-sysusers, possibly during
first boot, or automatically by rpm when the package is installed.

The existing polkitd_user config option is used for the user name.
The UID can be also set with polkitd_uid. This is useful for distributions
which want to use a statically allocated UID, like Fedora. It probably isn't needed
for anything, but let's add this possibility so that distros can transition to the
new mechanism more easily.

If the user already exists, sysusers will not do anything, to it's safe to add
this file on existing installations.

By having the sysusers file, creation of the user can be simplified in packages.
It is also easier to use polkit on systems after a "factory reset", i.e. the
removal of /etc: the user will be recreated automatically.